+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # CHANGELOG
 
+## v0.17.2
+
+- Fixed corruption roll enricher on danger roll chat card result
+
+## v0.17.1
+
+- Hotfix for Calendaria integration on Caravan sheet
+
+## v0.17.0
+
+- Added support to calendaria: SDM module
+
+## v0.16.0
+
+- Added support to many [text enrichers](https://github.com/luizrcb/foundryvtt-sdm/wiki/Text-Enrichers)
+- Added corruption roll enricher button to the failed danger roll chat card
+- Added defeat roll enricher button to the apply damage chat card when actor's current life is reduced to 0
+- Added burden item references to failed corruption rolls chat cards (you can drag and drop the burden item from the chat into a character sheet)
+- Added instructions about dropping actors on pet items to create a link
+
+## v0.15.0
+
+- Refactored Agents & Helpers tab on Caravan sheets
+- Added support to place Caravan Crew tokens on the scene canvas using the Place Crew Members button
+- Inventory roll icons were moved to item images on hover
+- When the usage of a power is dangerous, you can click the dangerous tag to perform a Danger Roll
+
+## v0.14.0
+
+- Added compact mode to Character and Caravan sheets
+
 ## v0.13.7
 
 - Added cash utility functions to Actors